Fix issue #23, about text alignment.
Setting the align property of text objects used to not only center-align
the text, but also reposition the object. This turned out to disrupt
the origin of the object. The align property now only sets the alignment,
and does not change the position of the object. Use the origin property
for that. This means the align property will only have an effect on
multi-line texts.

The baseline property can still be used, and will still reposition the object.