Major rewrite of the event system.
A bug regarding missing leave/enter events made me rethink how the
event model worked and led to a major rewrite of the whole event
system. It should now be easier to follow and be more stable.

Double tapping was also added to the touch module, with the event
name dbltap. The tap interval has a default of 500 ms, but can be
changed at runtime.