Another rewrite of the event system.
I realized the event system was still not behaving as one
would like, so I updated it to be even better. A flowchart
was created to set the flow right and the code now follows
that. This will make event propagation work, as well
as the stopPropagation method to stop the events from
bubbling up further up the parent chain.