Fix bug regarding pointer position.
When the pointer is moved outside the canvas, the position was set
to new coordinates. This was really noticable when the canvas was
in an iframe, and the pointer was moved out of the canvas and the
iframe at the same time. The position is no longer set when the
pointer is moved out of the canvas.