New structure for draw order. Fix issue #31.
The draw order was just in one level, which gave bad results when
adding new child objects to an existing object that was not on top.
The new child object would then be drawn on top of all other objects
and did not respect the parent's `zIndex`.

The new way uses the children list that every display object has.
This list was also added to the core instance for objects that are
added to the root level. When redrawing the canvas, it goes through
the children list of the core instance, starting from the end. For
each object it will recursively go through the children list until
no children are found and draw each one, which will draw them in
the correct order based on `zIndex`.

With this change, the `zIndex` will be local to the parent's children
list, not global. No matter the depth, if an object has three
children, the zIndexes will be 0, 1 and 2, where 2 is the front
object. The front object will then be `object.children[2]`.

The old one-level list `objects` in the draw module is now deprecated.
It is still there and will contain all currently added objects,
but please don't rely on it since it will be removed in a future
release.