Add new option to drag and drop: bubble
A new option can be set in the options object passed in to
dragAndDrop: bubble. If bubble is set to false, it will stop the
events from bubbling out to the parent chain. This makes it possible
to have drag and drop functionality on both an object and its
children, without them interfering with each other because of bubbling.