bugfix (#36): stop() and finish() for animations now work as intended.
The stop method in the animation module wasn't stopping the animations
correctly, because of a timer that was not cleared.

The final method didn't use the right queue object which caused errors
when trying to execute the callback function. It was also calling the
stop method, which suffered from the error above.