Rewrite animation module to use a more structured approach.
The method that handled the animations was getting way to big and
messy, so I rewrote it and split it out into different smaller
sections.

There is a new API for the animate method. The old syntax is still
supported, but marked as deprecated and will soon be removed. The
new syntax is simpler; you pass two simple objects to the animate
method on display objects:
displayObj.animate(properties, options);
The new thing is options, which is a way to simplify the different
options like duration, easing etc.

The logic is also made clearer by doing this. Every animation gets
its own animation object consisting of properties and options for
that animation.

To improve performance, the timer has now been changed to use
requestAnimationFrame.