Added ability for addChild and removeChild to add/remove multiple objects
This is done by placing all the objects you want to add/remove into an object, and then using that object as the displayobj variable in addChild or removeChild.

Both functions still support single objects. The code is a little messy, and I may look at cleaning it up later.