(patch) Respect cross-domain iframes
When adding mouse and touch events, it was looking for parent
windows to detect mouse movements there too. However, when the
iframe with the oCanvas content exists on a different domain
than the parent window, the checks would cause permission errors.
This avoids it by comparing the host of the referrer and the
current host (for the content inside the iframe).