(patch) displayobject/clone: Fix absolute positions for new object
Cloning an object will not give the new object the same parent as
the original object. The new will not have a parent. Because of
this, the new object must not have absolute positions that are
based on the original parent object. To solve this, I decrement
the positions by the absolute positions of the parent object.