Update index.js to include all modules
The modules are included in a structure that matches the folder
structure of the package. This makes it easy to change project
code from require statements to object access.

If the window object is available, requiring this file will put
an 'oCanvas' object on window, and it will have the same structure
as described above. The file outputted by QuickStart when Grunt
is run will automatically run this file and make everything
available on window.oCanvas.