Add canvas tests for classes/World
The canvas tests require the HTML Canvas API to be there. To make
that work in a node world (which mocha runs in), it requires the
npm package 'canvas' (node-canvas). That package can not be a dev
dependency, since it depends on system packages that can't be
installed via NPM. To get it working, follow these instructions:

https://github.com/LearnBoost/node-canvas#installation

Then run `mocha test/canvas`