classes/World: Make sure world property is restored
When restoring a world instance from JSON or a plain object, we
must make sure that the world property on cameras is set, to
also restore relationships. This happened before too, but was not
tested and used a mechanism in the fromObject method instead of
watching the collection, which is a more generic approach.