Add transformation matrix support
All canvas objects now have a new method: getTransformationMatrix
It will return a Matrix instance with the data of the combined
transformations for the object (translation, rotation and
scaling). It will also keep a caching system, where it uses a
cached matrix if it's still valid, and exposes a way to invalidate
the matrices.

The matrix calculations come from the npm package matrixmath
(written by koggdal).