CanvasObject: Fix failing tests
Some tests fail because the CanvasObject instances in the tests
don't update the cache properly. This is because the class lacks
implementation of a few methods, which subclasses (like
RectangularCanvasObject) should implement. Testing CanvasObject
in isolation requires mocking of a few methods, and so the mocked
methods need to manage the cache correctly.