Camera: Allow a reference to be passed to getTransformationMatrix
If a Canvas instance is passed to getTransformationMatrix, the
rotation will change sign (to adjust for that the canvas renders
what the camera sees, so the rendering should be rotated in the
opposite direction). The scaling should be the exact zoom value
if a Canvas instance is passed. If a Canvas instance is not
passed, the scaling should be the inverse of the zoom (1 / zoom),
and the rotation should be the normal rotation.