Add two utils: getClassName and isInstanceOf
The isInstanceOf utility can be used to check if an object is an
instance of a class with a certain name. The upside of this is
that the class itself is not needed, as long as we know the name.
By checking for name instead of the actual class reference, we can
avoid circular dependencies where classes are required in a module
only for doing a type check somewhere.

It uses the new utility getClassName to get the name of a
constructor function. That uses the className property if present
(on the constructor function), otherwise it checks
constructor.name (or fall back to parsing the stringified
function).