Make sure image and text are only redrawn if previously drawn
If a text object or an image object was created and added to
canvas without redrawing (using the redraw flag set to false),
the objects should not automatically redraw when the font or image
resource has loaded. They should only automatically redraw if the
object has been previously drawn (or tried to be drawn).