Bump the code coverage to 100% (including branch coverage)
Since the switch to Istanbul instead of Blanket, there is more
detailed code coverage with support for branch checking. As this
was not done before, the code coverage dropped from the usual 100%
to slightly less. Not a big problem, but slight changes to code or
a few more tests bumped it up to 100% again, for all types of
coverage.