Canvas: Draw bounding rectangles in the canvas coordinate space
Prior to fixing up the matrix and vertex methods, the bounding
rectangles were fetched in the World coordinate space (now renamed
to Scene). However, since we improved the signatures for these
methods so that they take a reference, we are now getting bounding
rectangles with coordinates relative to the canvas, not the scene.
This means we will not need to transform the canvas context to
the scene (we shouldn't, as that draws bounding rectangles in
wrong places).