utils/isInstanceOf: Slightly improve performance
Array.prototype.slice will create a new array object, which is not
needed when we can check with only strings. By using slice we are
also leaking the arguments object, which is not recommended:
https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#3-managing-arguments
Normally, this would not be a problem, but as this is a function
that is called a ton of times each frame, it needs to be optimized
as much as possible.