CanvasObject: Don't use arrays of matrices when multiplying
Using arrays will either create lots of new objects we don't need,
or we use a cache like we used to do here, but then you keep
changing the length of the array. The content of the array is not
really that dynamic, so we can just pass the matrices directly to
the multiply method instead of passing an array of matrices.