CanvasObject: Decrease multiplications needed in getPointIn
The desired outcome here is to get a matrix representing the
product of the transformation matrix and the input point matrix.
To do this, we had to create an identity matrix to start with
something that won't affect the outcome. However, with the new
version of the matrixmath package, we can just copy the data from
the transformation matrix into the output matrix, and then do a
single multiplication by the input point.