Camera: Invalidate obj cache when vertices reference matches
In some cases (when triggered from CanvasObject#getPointIn) the
reference passed to CanvasObject#getTransformationMatrix is not
the original reference object, but an object one step in (for
example, if the reference passed to getPointIn is a canvas, the
reference it will then pass to getTransformationMatrix is the
scene).

For this reason, we can't always check against that reference when
invalidating the camera cache. We must also look at a cache for
the original reference. The one that calls getPointIn is
getVertices, so we can use the cache unit vertices-reference.