Add new functionality: Tracker
The tracker functionality allows objects or cameras to track the
position of other objects or cameras. There is support for common
techniques used for camera movement in games like only tracking
inside certain boundaries (game map) and only move camera when the
target (player) moves outside a certain window. It also supports
influencers (attractors, repellers and pushers) that will affect
the target position of the object. These objects can be placed in
a game map for example to draw attention for the camera to a
certain part of the map.

Example:
var tracker = new Tracker({
  object: camera,
  target: player
});
tracker.start();