Make Circle inherit from Ellipse instead of EllipticalCanvasObject
As a circle is just one form of ellipse, it can be rendered
properly using the Ellipse class. That means Circle can inherit
directly from Ellipse instead of the base class
EllipticalCanvasObject. The only thing that is different with
Circle is now that it has a `radius` property that is tied to
`radiusX` and `radiusY`.