CanvasObject: Add support for rendering fills and strokes
Rendering fills and strokes is quite massive, as a fill can be
a color, different textures or a camera. To make it easier to
apply all these fills to a shape class, the base for it will be
in CanvasObject, and methods specific to rendering a certain
type of fill can be defined in the subclass.