build: Split into two modes, dev and release
As the built files for release are now checked in to the repo, the
default build command should not build files to that folder. It will
now build files to build/dev/ (which is now in .gitignore). When
building for a release, it will place the files in build/dist/.

This will put files in build/dev/
node build/build.js

This will put files in build/dist/
node build/build.js release