sprite: Fix loading of images
This was previously fixed for the image display object, but not for
sprites. This commit also checks if the provided image element is
already loaded, in which case it doesn't create a new image to load.
That helps if using an asset manager, as it then makes no sense at
all to wait for images to load after the asset manager has loaded
them.

Earlier fix for image display object:
cf8d8b05fb76a9967f1a88e3e63e7b4cb170b218