displayobjects: Fix issue with shadows
Internally the shadow data is stored in an object, which happened to get shared
across all display objects, instead of being unique to each object.

This would ideally be fixed in another more central place, but this approach
keeps most compatibility. For example, making sure `oCanvas.extend` handles
nested data structures could be a breaking change for some apps, if they relied
on the behavior of only extending shallowly.